LineStringと設備の表示切替を追加
